Common train_exact_module
